+++ castero/perspective.py
@@ -1,3 +1,5 @@
+import curses
+import textwrap
 from abc import ABC, abstractmethod
 
 from castero.menu import Menu
@@ -80,6 +82,11 @@ class Perspective(ABC):
         """
 
     @abstractmethod
+    def made_active(self) -> None:
+        """Called each time the perspective is made active (switched to).
+        """
+
+    @abstractmethod
     def _get_active_menu(self) -> Menu:
         """Retrieve the active Menu, if there is one.
 
@@ -109,3 +116,121 @@ class Perspective(ABC):
         new_active_menu = self._get_active_menu()
         if new_active_menu is not None:
             new_active_menu.set_active(True)
+
+    def _draw_metadata(self, window, feed=None, episode=None) -> None:
+        """Draws the metadata of the selected feed/episode onto the window.
+
+        Exactly one of feed or episode must be specified.
+
+        Args:
+            window: the curses window which will display the metadata
+            feed: (optional) the Feed whose metadata will be displayed
+            episode: (optional) the Episode whose metadata will be displayed
+        """
+        assert window is not None
+        assert feed != episode and (feed is None or episode is None)
+
+        output_lines = []  # 2D array, each element is [attr, str]
+        max_lines = window.getmaxyx()[0] - 2
+        max_line_width = window.getmaxyx()[1] - 1
+
+        # clear the window by drawing blank lines
+        for y in range(2, window.getmaxyx()[0]):
+            window.addstr(y, 0, " " * max_line_width)
+
+        if feed is not None:
+            # draw feed title
+            self._append_metadata_lines(window, feed.title, output_lines,
+                                        attr=curses.A_BOLD)
+            # draw feed lastBuildDate
+            self._append_metadata_lines(window, feed.last_build_date,
+                                        output_lines,
+                                        add_blank=True)
+            # draw feed link
+            self._append_metadata_lines(window, feed.link, output_lines,
+                                        add_blank=True)
+            # draw feed description
+            self._append_metadata_lines(window, "Description:", output_lines,
+                                        attr=curses.A_BOLD)
+            self._append_metadata_lines(window, feed.description, output_lines,
+                                        add_blank=True)
+            # draw feed copyright
+            self._append_metadata_lines(window, "Copyright:", output_lines,
+                                        attr=curses.A_BOLD)
+            self._append_metadata_lines(window, feed.copyright, output_lines,
+                                        add_blank=True)
+            # draw feed number of episodes
+            num_dl = sum([episode.downloaded(self._display.config) for
+                          episode in feed.episodes])
+            self._append_metadata_lines(window, "Episodes:", output_lines,
+                                        attr=curses.A_BOLD)
+            self._append_metadata_lines(window,
+                                        "Found %d episodes (%d downloaded)" % (
+                                            len(feed.episodes), num_dl
+                                        ), output_lines
+                                        )
+        elif episode is not None:
+            # draw episode title
+            self._append_metadata_lines(window, episode.title, output_lines,
+                                        attr=curses.A_BOLD)
+            # draw episode pubdate
+            self._append_metadata_lines(window, episode.pubdate, output_lines,
+                                        add_blank=True)
+            # draw episode link
+            self._append_metadata_lines(window, episode.link, output_lines,
+                                        add_blank=True)
+            # draw episode description
+            self._append_metadata_lines(window, "Description:", output_lines,
+                                        attr=curses.A_BOLD)
+            self._append_metadata_lines(window, episode.description,
+                                        output_lines,
+                                        add_blank=True)
+            # draw episode copyright
+            self._append_metadata_lines(window, "Copyright:", output_lines,
+                                        attr=curses.A_BOLD)
+            self._append_metadata_lines(window, episode.copyright,
+                                        output_lines,
+                                        add_blank=True)
+
+            # draw episode downloaded
+            self._append_metadata_lines(window, "Downloaded:", output_lines,
+                                        attr=curses.A_BOLD)
+            self._append_metadata_lines(window,
+                                        "Episode downloaded and available for offline playback." if
+                                        episode.downloaded(
+                                            self._display.config) else
+                                        "Episode not downloaded.",
+                                        output_lines)
+
+        y = 2
+        for line in output_lines[:max_lines]:
+            window.attrset(curses.color_pair(1))
+            if line[0] != -1:
+                window.attron(line[0])
+            window.addstr(y, 0, line[1])
+            y += 1 + line[1].count('\n')
+
+    def _append_metadata_lines(self, window, string, output_lines, attr=-1,
+                               add_blank=False) -> None:
+        """Appends properly formatted lines to the 2D output_lines array.
+
+        Args:
+            window: the curses window which will display the metadata
+            string: the string to add to output_lines
+            output_lines: 2D array, each element is [attr, str]
+            attr: (optional) the attribute (i.e. curses.A_BOLD)
+        """
+        max_lines = int(0.7 * window.getmaxyx()[0])
+        max_line_width = window.getmaxyx()[1] - 1
+        lines = textwrap.wrap(string, max_line_width)
+
+        # truncate to at most 70% of the total lines on the screen
+        lines = lines[:max_lines]
+
+        # add all lines to array
+        for line in lines:
+            output_lines.append([attr, line])
+
+        # add a blank line afterward, if necessary
+        if add_blank:
+            output_lines.append([-1, ""])