+++ castero/perspectives/primary.py
@@ -1,5 +1,4 @@
 import curses
-import textwrap
 
 from castero import helpers
 from castero.menu import Menu
@@ -130,7 +129,15 @@ class Primary(Perspective):
 
         # draw metadata
         if not self._metadata_updated:
-            self._draw_metadata()
+            selected_feed_index = self._feed_menu.selected_index
+            feed = self._display.feeds.at(selected_feed_index)
+            if feed is not None:
+                if self._active_window == 0:
+                    self._draw_metadata(self._metadata_window, feed=feed)
+                elif self._active_window == 1:
+                    selected_episode_index = self._episode_menu.selected_index
+                    episode = feed.episodes[selected_episode_index]
+                    self._draw_metadata(self._metadata_window, episode=episode)
 
     def handle_input(self, c) -> bool:
         """Performs action corresponding to the user's input.
@@ -170,7 +177,6 @@ class Primary(Perspective):
             queue.clear()
             self._create_player_from_selected()
             queue.play()
-            self._get_active_menu().move(-1)
         elif c == key_mapping[config['key_add_selected']]:
             self._create_player_from_selected()
             self._get_active_menu().move(-1)
@@ -207,6 +213,12 @@ class Primary(Perspective):
 
         return keep_running
 
+    def made_active(self) -> None:
+        """Called each time the perspective is made active (switched to).
+
+        Overrides method from Perspective; see documentation in that class.
+        """
+
     def refresh(self) -> None:
         """Refresh the screen and all windows.
 
@@ -228,124 +240,6 @@ class Primary(Perspective):
             1: self._episode_menu,
         }.get(self._active_window)
 
-    def _draw_metadata(self) -> None:
-        """Draws the metadata of the selected feed/episode onto the window.
-        """
-        assert self._metadata_window is not None
-
-        output_lines = []  # 2D array, each element is [attr, str]
-        max_lines = self._metadata_window.getmaxyx()[0] - 2
-        max_line_width = self._metadata_window.getmaxyx()[1] - 1
-
-        # clear the window by drawing blank lines
-        for y in range(2, self._metadata_window.getmaxyx()[0]):
-            self._metadata_window.addstr(y, 0, " " * max_line_width)
-
-        if self._active_window == 0:
-            # * the selected item is a feed
-            selected_index = self._feed_menu.selected_index
-            feed = self._display.feeds.at(selected_index)
-
-            if feed is not None:
-                # draw feed title
-                self._append_metadata_lines(feed.title, output_lines,
-                                            attr=curses.A_BOLD)
-                # draw feed lastBuildDate
-                self._append_metadata_lines(feed.last_build_date, output_lines,
-                                            add_blank=True)
-                # draw feed link
-                self._append_metadata_lines(feed.link, output_lines,
-                                            add_blank=True)
-                # draw feed description
-                self._append_metadata_lines("Description:", output_lines,
-                                            attr=curses.A_BOLD)
-                self._append_metadata_lines(feed.description, output_lines,
-                                            add_blank=True)
-                # draw feed copyright
-                self._append_metadata_lines("Copyright:", output_lines,
-                                            attr=curses.A_BOLD)
-                self._append_metadata_lines(feed.copyright, output_lines,
-                                            add_blank=True)
-                # draw feed number of episodes
-                num_dl = sum([episode.downloaded(self._display.config) for
-                              episode in feed.episodes])
-                self._append_metadata_lines("Episodes:", output_lines,
-                                            attr=curses.A_BOLD)
-                self._append_metadata_lines(
-                    "Found %d episodes (%d downloaded)" % (
-                        len(feed.episodes), num_dl
-                    ), output_lines
-                )
-
-        elif self._active_window == 1:
-            # * the selected item is an episode
-            selected_feed_index = self._feed_menu.selected_index
-            selected_episode_index = self._episode_menu.selected_index
-            feed = self._display.feeds.at(selected_feed_index)
-
-            if feed is not None:
-                episode = feed.episodes[selected_episode_index]
-
-                # draw episode title
-                self._append_metadata_lines(episode.title, output_lines,
-                                            attr=curses.A_BOLD)
-                # draw episode pubdate
-                self._append_metadata_lines(episode.pubdate, output_lines,
-                                            add_blank=True)
-                # draw episode link
-                self._append_metadata_lines(episode.link, output_lines,
-                                            add_blank=True)
-                # draw episode description
-                self._append_metadata_lines("Description:", output_lines,
-                                            attr=curses.A_BOLD)
-                self._append_metadata_lines(episode.description, output_lines,
-                                            add_blank=True)
-                # draw episode copyright
-                self._append_metadata_lines("Copyright:", output_lines,
-                                            attr=curses.A_BOLD)
-                self._append_metadata_lines(episode.copyright, output_lines,
-                                            add_blank=True)
-
-                # draw episode downloaded
-                self._append_metadata_lines("Downloaded:", output_lines,
-                                            attr=curses.A_BOLD)
-                self._append_metadata_lines(
-                    "Episode downloaded and available for offline playback." if
-                    episode.downloaded(self._display.config) else
-                    "Episode not downloaded.", output_lines)
-
-        y = 2
-        for line in output_lines[:max_lines]:
-            self._metadata_window.attrset(curses.color_pair(1))
-            if line[0] != -1:
-                self._metadata_window.attron(line[0])
-            self._metadata_window.addstr(y, 0, line[1])
-            y += 1 + line[1].count('\n')
-
-    def _append_metadata_lines(self, string, output_lines, attr=-1,
-                               add_blank=False) -> None:
-        """Appends properly formatted lines to the 2D output_lines array.
-
-        Args:
-            string: the string to add to output_lines
-            output_lines: 2D array, each element is [attr, str]
-            attr: (optional) the attribute (i.e. curses.A_BOLD)
-        """
-        max_lines = int(0.7 * self._metadata_window.getmaxyx()[0])
-        max_line_width = self._metadata_window.getmaxyx()[1] - 1
-        lines = textwrap.wrap(string, max_line_width)
-
-        # truncate to at most 70% of the total lines on the screen
-        lines = lines[:max_lines]
-
-        # add all lines to array
-        for line in lines:
-            output_lines.append([attr, line])
-
-        # add a blank line afterward, if necessary
-        if add_blank:
-            output_lines.append([-1, ""])
-
     def _create_player_from_selected(self) -> None:
         """Creates player(s) based on the selected items and adds to the queue.
 
@@ -361,13 +255,14 @@ class Primary(Perspective):
         if self._active_window == 0:
             if feed is not None:
                 for episode in feed.episodes:
-                    player = Player(str(episode), episode.get_playable())
+                    player = Player(str(episode), episode.get_playable(),
+                                    episode)
                     self._display.queue.add(player)
         elif self._active_window == 1:
             episode_index = self._episode_menu.selected_index
             if feed is not None:
                 episode = feed.episodes[episode_index]
-                player = Player(str(episode), episode.get_playable())
+                player = Player(str(episode), episode.get_playable(), episode)
                 self._display.queue.add(player)
 
     def _invert_selected_menu(self) -> None: